Make Board component
This was an interesting experiment. I started by moving all of the
code so that Board exposed the model/view/update signature
of the Elm architecture and then used the compiler/type system
to make sure everything was wired together properly.

Once I got it to pass again, everything just worked.

There is still some 'magic' that is going on with Signal.forwardTo
and Effects.map but I will figure those out later.